# openapi/paths.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:4000
...
paths:
...
  /tweets:
    post:
      tags:
        - tweets
      summary: Tạo tweet
      description: Tạo tweet
      operationId: createTweet
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin tạo tweet
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TweetRequestBody'
        required: true
      responses:
        default:
          description: Tạo tweet thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Tạo tweet thành công
    get:
      tags:
        - tweets
      summary: Lấy danh sách tweet từ những người mình theo dõi
      description: Lấy danh sách tweet từ những người mình theo dõi
      operationId: getTweets
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Số trang
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: Số lượng tweet trên 1 trang
          required: false
          schema:
            type: integer
            example: 10
      responses:
        default:
          description: Lấy danh sách tweet thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Lấy danh sách tweet thành công
                  result:
                    type: object
                    properties:
                      tweets:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tweet'
                      total_page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      page:
                        type: integer
                        example: 1
...
  /likes:
    post:
      tags:
        - likes
      summary: Like tweet
      description: Like tweet
      operationId: likeTweet
      security:
        - BearerAuth: []
      requestBody:
        description: Tweet ID
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequestBody'
        required: true
      responses:
        default:
          description: Like tweet successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: LIKE Tweet successfully
                  result:
                    $ref: '#/components/schemas/UnTweetResponseBody'
    get:
      tags:
        - likes
      summary: Get Tweets User Liked
      description: Get Tweets User Liked
      operationId: tweetUserLiked
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Số trang
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: Số lượng tweet trên 1 trang
          required: false
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Get tweets liked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Get tweets liked successfully
                  result:
                    type: object
                    properties:
                      tweets:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tweet'
                      total_page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      page:
                        type: integer
                        example: 1
        '400':
          description: Invalid status value
  /likes/tweets/{tweet_id}:
    delete:
      tags:
        - likes
      summary: Unlike tweet by tweet_id
      description: Unlike tweet by tweet_id
      operationId: unlikeTweet
      security:
        - BearerAuth: []
      parameters:
        - name: tweet_id
          in: path
          description: Tweet Id
          required: true
          schema:
            type: string
            format: MongoId
            example: 64be0ad2e43d2464394feedb
      responses:
        default:
          description: Unlike successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unlike successfully
                  result:
                    $ref: '#/components/schemas/UnTweetResponseBody'
  /likes/{like_id}:
    delete:
      tags:
        - likes
      summary: Unlike tweet by like_id
      description: Unlike tweet by like_id
      operationId: unlikeTweet
      security:
        - BearerAuth: []
      parameters:
        - name: like_id
          in: path
          description: Like Id
          required: true
          schema:
            type: string
            format: MongoId
            example: 64be0ad2e43d2464394feedb
      responses:
        default:
          description: Unlike successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unlike successfully
                  result:
                    $ref: '#/components/schemas/UnTweetResponseBody'
  /bookmarks:
    post:
      tags:
        - bookmarks
      summary: Bookmark tweet
      description: Bookmark tweet
      operationId: bookmarkTweet
      security:
        - BearerAuth: []
      requestBody:
        description: Tweet ID
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkRequestBody'
        required: true
      responses:
        default:
          description: Bookmark tweet successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Bookmark Tweet successfully
                  result:
                    $ref: '#/components/schemas/UnTweetResponseBody'
    get:
      tags:
        - bookmarks
      summary: Get Tweets User Bookmarked
      description: Get Tweets User Bookmarked
      operationId: tweetUserBookmarked
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Số trang
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: Số lượng tweet trên 1 trang
          required: false
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Get tweets bookmarked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Get tweets bookmarked successfully
                  result:
                    type: object
                    properties:
                      tweets:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tweet'
                      total_page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      page:
                        type: integer
                        example: 1
        '400':
          description: Invalid status value
  /bookmarks/tweets/{tweet_id}:
    delete:
      tags:
        - bookmarks
      summary: Unbookmark tweet by tweet_id
      description: Unbookmark tweet by tweet_id
      operationId: unbookmarkTweet
      security:
        - BearerAuth: []
      parameters:
        - name: tweet_id
          in: path
          description: Tweet Id
          required: true
          schema:
            type: string
            format: MongoId
            example: 64be0ad2e43d2464394feedb
      responses:
        default:
          description: Unbookmark successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unbookmark successfully
                  result:
                    $ref: '#/components/schemas/UnTweetResponseBody'
  /bookmarks/{bookmark_id}:
    delete:
      tags:
        - bookmarks
      summary: Unbookmark tweet by like_id
      description: Unbookmark tweet by like_id
      operationId: unbookmarkTweet
      security:
        - BearerAuth: []
      parameters:
        - name: bookmark_id
          in: path
          description: Bookmark Id
          required: true
          schema:
            type: string
            format: MongoId
            example: 64be0ad2e43d2464394feedb
      responses:
        default:
          description: Unbookmark successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unlike successfully
                  result:
                    $ref: '#/components/schemas/UnTweetResponseBody'
...